refactor(capabilities): split component into a submodule
Closes #2168
